integration Rafactoring, add warehousing module

+++ app/internal/catalog/box/section/adm/controller/httpv1/box_model.go
@@ -32,6 +32,6 @@ type (
 	// BoxListResponse - comment struct.
 	BoxListResponse struct {
 		Items []entity.Box `json:"items"`
-		Total uint64       `json:"total"`
+		Total int          `json:"total"`
 	}
 )
