+++ app/internal/dictionaries/paperfacture/section/adm/controller/httpv1/paper_facture_model.go
@@ -19,6 +19,6 @@ type (
 	// PaperFactureListResponse - comment struct.
 	PaperFactureListResponse struct {
 		Items []entity.PaperFacture `json:"items"`
-		Total uint64                `json:"total"`
+		Total int                   `json:"total"`
 	}
 )
