+++ app/internal/dictionaries/printformat/section/adm/controller/httpv1/print_format_model.go
@@ -24,6 +24,6 @@ type (
 	// PrintFormatListResponse - comment struct.
 	PrintFormatListResponse struct {
 		Items []entity.PrintFormat `json:"items"`
-		Total uint64               `json:"total"`
+		Total int                  `json:"total"`
 	}
 )
