+++ app/internal/factory/catalog/box/section/adm/unit_box.go
@@ -21,7 +21,7 @@ func initBoxController(
 	dbConnManager mrstorage.DBConnManager,
 	requestExtendParser *validate.ExtendParser,
 	responseSender mrserver.ResponseSender,
-	pageSizeMax uint64,
+	pageSizeMax int,
 ) (mrserver.HttpController, error) {
 	entityMeta, err := mrsql.ParseEntity(logger, entity.Box{})
 	if err != nil {
