+++ app/internal/provideraccounts/section/adm/company_page.go
@@ -9,11 +9,11 @@ import (
 type (
 	// CompanyPageUseCase - comment interface.
 	CompanyPageUseCase interface {
-		GetList(ctx context.Context, params entity.CompanyPageParams) (items []entity.CompanyPage, countItems uint64, err error)
+		GetList(ctx context.Context, params entity.CompanyPageParams) (items []entity.CompanyPage, countItems int, err error)
 	}
 
 	// CompanyPageStorage - comment interface.
 	CompanyPageStorage interface {
-		FetchWithTotal(ctx context.Context, params entity.CompanyPageParams) (rows []entity.CompanyPage, countRows uint64, err error)
+		FetchWithTotal(ctx context.Context, params entity.CompanyPageParams) (rows []entity.CompanyPage, countRows int, err error)
 	}
 )
